requirements.txt
torch>=1.9.0
torchvision>=0.10.0
albumentations>=1.0.0
segmentation-models-pytorch>=0.2.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0
tqdm>=4.60.0